chore: moved tests to the relatated folder
