+++ src/components/AppLayout/Header/__tests__/Header.spec.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { Header } from "../Header";
+import { Header } from "..";
 
 const testText = "This is a test content.";
 const mockFn = jest.fn();
